requirements.txt
python-telegram-bot
moviepy
numpy
opencv-python
